Switch to Babel & ES6 in Gulp for React static build
- Rename gulpfile.js -> gulpfile.babel.js to enable ES6 via Babel

- Move generate-static-html.js functions into frontend/tasks/pages.js
  and import pages JS now that we can run them directly without a
  pre-build step

- Delete Gulp tasks scripts-generate-static-html and
  scripts-app-individual since we don't need those for an ES6 build
  process

- Move frontend/src/app.js back to frontend/src/app/index.js

- Inject data-page-name & data-page-param attributes into <body> to
  drive app bootstrap instead of window location routing.
