+++ gulpfile.babel.js
@@ -65,13 +65,13 @@ gulp.task('default', done => runSequence(
 
 // Exit if the gulpfile changes so we can self-reload with a wrapper script.
 gulp.task('self-watch', () => gulp.watch([
-  './gulpfile.js',
+  './gulpfile.babel.js',
   './frontend/config.js',
   './debug-config.json',
   './frontend/tasks/*.js'
 ], () => process.exit()));
 
-gulp.task('self-lint', () => gulp.src('gulpfile.js')
+gulp.task('self-lint', () => gulp.src('gulpfile.babel.js')
   .pipe(eslint())
   .pipe(eslint.format())
   .pipe(eslint.failOnError()));
